抽取公共API
