--- rx-wecom-sdk/src/main/java/cn/felord/reactive/SuiteDetail.java
@@ -1,35 +1,0 @@
-/*
- *  Copyright (c) 2023. felord.cn
- *    Licensed under the Apache License, Version 2.0 (the "License");
- *    you may not use this file except in compliance with the License.
- *  You may obtain a copy of the License at
- *       https://www.apache.org/licenses/LICENSE-2.0
- *  Website:
- *       https://felord.cn
- *  Unless required by applicable law or agreed to in writing, software
- *  distributed under the License is distributed on an "AS IS" BASIS,
- *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  See the License for the specific language governing permissions and
- *  limitations under the License.
- */
-
-package cn.felord.reactive;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-/**
- * @author dax
- * @since 2021/12/8 10:08
- */
-@AllArgsConstructor
-@Getter
-public class SuiteDetail {
-    private final String suiteId;
-    private final String suiteSecret;
-    private String suiteTicket;
-
-    void setSuiteTicket(String suiteTicket) {
-        this.suiteTicket = suiteTicket;
-    }
-}
