+++ rx-wecom-sdk/src/main/java/cn/felord/reactive/api/WorkWeChatApiClient.java
@@ -16,7 +16,7 @@
 package cn.felord.reactive.api;
 
 
-import cn.felord.reactive.AgentDetails;
+import cn.felord.AgentDetails;
 import cn.felord.reactive.RetrofitFactory;
 import okhttp3.ConnectionPool;
 import okhttp3.logging.HttpLoggingInterceptor;
