+++ wecom-objects/src/main/java/cn/felord/AgentDetails.java
