+++ wecom-objects/src/main/java/cn/felord/SuiteDetail.java
