+++ wecom-objects/src/main/java/cn/felord/WeComAgentTicketCacheable.java
@@ -1,21 +1,19 @@
 /*
- * Copyright (c) 2023. felord.cn
- *   Licensed under the Apache License, Version 2.0 (the "License");
- *   you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *      https://www.apache.org/licenses/LICENSE-2.0
- * Website:
- *      https://felord.cn
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
+ *  Copyright (c) 2023. felord.cn
+ *    Licensed under the Apache License, Version 2.0 (the "License");
+ *    you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *       https://www.apache.org/licenses/LICENSE-2.0
+ *  Website:
+ *       https://felord.cn
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
  */
 
 package cn.felord;
-
-import org.springframework.lang.NonNull;
 
 /**
  * Ticket 缓存
@@ -34,7 +32,7 @@ public interface WeComAgentTicketCacheable extends WeComCacheable {
      * @param corpTicket the corpTicket
      * @return the corpTicket
      */
-    String putCorpTicket(@NonNull String corpId, @NonNull String agentId, @NonNull String corpTicket);
+    String putCorpTicket(String corpId, String agentId, String corpTicket);
 
     /**
      * 获取企业的jsapi_ticket.
@@ -43,7 +41,7 @@ public interface WeComAgentTicketCacheable extends WeComCacheable {
      * @param agentId the agentId
      * @return the corpTicket
      */
-    String getCorpTicket(@NonNull String corpId, @NonNull String agentId);
+    String getCorpTicket(String corpId, String agentId);
 
     /**
      * 放入应用的jsapi_ticket.
@@ -55,7 +53,7 @@ public interface WeComAgentTicketCacheable extends WeComCacheable {
      * @param agentTicket the agentTicket
      * @return the agentTicket
      */
-    String putAgentTicket(@NonNull String corpId, @NonNull String agentId, @NonNull String agentTicket);
+    String putAgentTicket(String corpId, String agentId, String agentTicket);
 
     /**
      * 获取应用的jsapi_ticket.
@@ -64,6 +62,6 @@ public interface WeComAgentTicketCacheable extends WeComCacheable {
      * @param agentId the agentId
      * @return the agentTicket
      */
-    String getAgentTicket(@NonNull String corpId, @NonNull String agentId);
+    String getAgentTicket(String corpId, String agentId);
 
 }
